refactor: move constants to `shared`
