+++ compiler/fs.ts
@@ -1,6 +1,6 @@
 import { dirname } from 'path';
-import { SEDGE_NAMESPACE } from '../core/constants.ts';
 import { Result } from '../core/types.ts';
+import { SEDGE_NAMESPACE } from '../shared/constants.ts';
 import { toExtension } from './path.ts';
 
 export interface SedgeFileSystem {
